Expand use of proxmox_api_validate_cert
The collection makes use of proxmox and uri Ansible modules.

By default, the proxmox modules **do not** validate the SSL
certificates, while the `ansible.builtin.uri` module **does** validate
the SSL certificates, which can create a discrepancy in the
configuration.

Expand the use of the variable `proxmox_api_validate_cert` to all the
roles of the collection, and use it for every module that support
validation of SSL certificates. This variable can now be used to
enable/disable the SSL certificate validation in uri and proxmox
modules.

Define the default value for `proxmox_api_validate_certs` to `false`.