feat: Allow to update VMs with new values
Because the proxmox_kvm module is not idempotent when `update: true`,
add a new variable `proxmox_kvm_update` that allows to run the update on
demand.

Also replace `refresh` with `update` which is more accurate and matches
the name of the module parameter. Note that `item.update` is not usable
since `update` is a method of python dicts. Ensure that `update` is
defined before using it.

Closes: #30
Fixes: #29