fix: Update containers in a new task
There is an undocumented rule which defines that ostemplate and update
are mutually exclusive. Add a new task that will update the containers
if required.

This duplicates the whole task, which could lead to some maintenance
issues in the future. However, trying to keep a single task with `omit`
for ostemplate when update is required did not allow to bypass the
condition of mutual exclusion.

Fix #36